Don't go through SupertypesLoopChecker for well-known types
Supertypes of ClassTypeConstructorImpl are known at the moment of
instantiation (i.e. it is never instantiated with some supertypes
coming from user's code) -- hence it is redundant to go through
loop disconnection in AbstractTypeConstructor.

The main point of this change is to fix race condition in some cases:
ClassTypeConstructorImpl is created with
LockBasedStorageManager.NO_LOCKS, and lazy values with post compute are
thread-unsafe in such case. Notably, other threads can observe
uninitialized supertypes on built-in types like Serializable/Cloneable.