Introduce UnsafeClassTypeConstructor
It is a TypeConstructor with known (and trusted) supertypes, and it
doesn't perform loop disconnection for them.

Use it ClassConstructorImpl instead of ClassTypeConstructorImpl -- it is
safe, because ClassConstructorImpl is created only for well-known types
(like Serializable, Cloneable, etc.)
Not that we can't change ClassTypeConstructorImpl -- its usages in
ClassConstructorImpl are safe indeed, but there are others (notably, in
LazySubstitutingDecscriptor).

This is a temporary measure to fix races on supertypes for Serializable
and Cloneable: previously, they were created with
LockBasedStorageManager.NO_LOCKS, which is totally unsafe for
AbstractTypeConstructor, and could lead to other threads observing
uninitialized supertypes.